fix cross building of sql server
in https://github.com/wal-g/wal-g/pull/1866 were introduced "linux only"
signal handling.  that signal handling makes some sense for PG (however,
I've no idea). no sense outside PG, but it also breaks cross building

export GOOS=windows
go get -v -t -d ./...
make deps
make sqlserver_build

...

 github.com/wal-g/wal-g/internal/config
../../internal/config/config.go:814:31: undefined: syscall.SIGUSR1
